Add SubViews approach and first SubView - Dashboard
Moved Urls.js to a separate module since now it deserves a better home
:-) In addition to that, JiraItem.qml does not have ready signal anymore
and instead of that, introduced a property with the same name - use case
remains the same, but now an approach is more flexible.
First SubView does not have any visual representations, but writes a
warning message with a current Jira instance.