Polish code style
Besides code style polishing, a Jira server input field will be updated
with an already selected Jira server as a part of the future "last mode"
handling. YesNo popup closing will be handled by generic popup. Usage of
JS "this" was removed in order to avoid implicit mistakes in the future.