Control visibility of tools based on current user permissions
After successful validation of Jira server, a request for currently
available permissions will be sent - it is necessary for visibility of
"Go Anonymous" button on Credentials Input page. Normally, at least
Browse permission is necessary for few of tools - if there are 0
permissions, this button will be hidden. Several other permissions are
verified after succssful user authentication and based on these
permissions, some tools would be either visible or not.